py: Fix WebKitGTK driver name.
Similar to WPEWebKit changes from 24d88d73636e33, WebKitGTK does not
follow the simple capitalization scheme for the driver name.

This commit also changes the default `WebKitWebDriver` binary to the
first one in `PATH` through shutil.which, instead of just the binary
name.

(Note: The effort to move to `pathlib` instead of `shutil`[1] is more
related to shutil's functions that work on files and directories, like
copy and move, instead of `which`, which just searches the executable
paths for the given command.)

[1] https://discuss.python.org/t/incrementally-move-high-level-path-operations-from-shutil-to-pathlib/19208