Using existing logger functionality.
Messages will be shown to the console
using the Logger functionality.

Adding a test that listens to the
logger and checks that logs are being
sent to console.

Fixing an issue when sending the log
to a file, as the OnDriverProcessStarted
method in the base class was not being
invoked.